Update NoteCard and NoteDetails components, and handle user authentication in tasks module

+++ convex/tasks.ts
@@ -1,23 +1,24 @@
-import { mutation, query } from "./_generated/server";
-import { ConvexError, v } from "convex/values";
+import { MutationCtx, QueryCtx, mutation, query } from "./_generated/server";
+import { v } from "convex/values";
 
-export const get = query({
-  args: {},
-  handler: async (ctx) => {
-    return await ctx.db.query("tasks").collect();
-  },
-});
+export async function handleAccess(ctx: QueryCtx | MutationCtx) {
+  const identity = await ctx.auth.getUserIdentity();
+  if (!identity) {
+    return null;
+  }
+
+  return identity;
+}
 
 export const addTask = mutation({
   args: { title: v.string(), description: v.string(), userId: v.string() },
   handler: async (ctx, { title, description, userId }) => {
-    const identity = await ctx.auth.getUserIdentity();
-    if (identity === null) {
-      throw new ConvexError("User not authenticated");
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
     }
-    if (identity.subject !== userId) {
-      throw new ConvexError("User not authorized");
-    }
+
     const task = await ctx.db.insert("tasks", {
       title,
       description,
@@ -31,6 +32,11 @@ export const addTask = mutation({
 export const deleteTask = mutation({
   args: { id: v.id("tasks") },
   handler: async (ctx, args) => {
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
+    }
     await ctx.db.delete(args.id);
   },
 });
@@ -38,6 +44,12 @@ export const deleteTask = mutation({
 export const toggleCompleteTask = mutation({
   args: { id: v.id("tasks") },
   handler: async (ctx, { id }) => {
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
+    }
+
     const task = await ctx.db.get(id);
     if (!task) {
       throw new Error(`No task found with id ${id}`);
@@ -49,6 +61,11 @@ export const toggleCompleteTask = mutation({
 export const updateTask = mutation({
   args: { id: v.id("tasks"), title: v.string(), description: v.string() },
   handler: async (ctx, { id, title, description }) => {
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
+    }
     const task = await ctx.db.get(id);
     if (!task) {
       throw new Error(`No task found with id ${id}`);
@@ -60,6 +77,12 @@ export const updateTask = mutation({
 export const getTaskList = query({
   args: { userId: v.string() },
   handler: async (ctx, args) => {
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
+    }
+
     const tasks = await ctx.db
       .query("tasks")
       .withIndex("by_userId")
@@ -73,6 +96,12 @@ export const getTaskList = query({
 export const getTask = query({
   args: { id: v.id("tasks") },
   handler: async (ctx, args) => {
+    const hasAccess = await handleAccess(ctx);
+
+    if (!hasAccess) {
+      return null;
+    }
+
     const task = await ctx.db.get(args.id);
     if (!task) {
       throw new Error(`No task found with id ${args.id}`);
